[#45] Refactor: WKWebView extension 분리
